+++ db/migrate/20260830230000_add_groups_refresh_run_id_to_access_token_details.rb
@@ -1,0 +1,5 @@
+class AddGroupsRefreshRunIdToAccessTokenDetails < ActiveRecord::Migration[8.0]
+  def change
+    add_column :access_token_details, :groups_refresh_run_id, :uuid
+  end
+end